Fixing complexity issues.
